third_party/blink/web_tests/http/tests/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-subframe-expected.txt

CONSOLE MESSAGE: PASS: Exception thrown successfully.
The scenario for this test is that you have an iframe with content from a foreign domain. In that foreign content is an iframe which loads a data: URL. This tests that the data: URL loaded iframe does not have access to the main frame using top.document.

Pass: Cross frame access from a data: URL on a different domain was denied.

--------
Frame: 'aFrame'
--------
Inner iframe on a foreign domain.

--------
Frame: '<!--framePath //aFrame/<!--frame0-->-->'
--------
Inner-inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame attempting to access the main frame. It should not have access to it.
